Create Timer to generate subscriptions' reports automatically
