+++ server/.env.example
@@ -1,4 +1,5 @@
 PORT=
+TIMER = 
 
 TWITTER_APP_API_KEY=
 TWITTER_APP_API_SECRET=
